Fixed DenseNet
The old one was referencing the DenseNet for image classification, not the semantic segmentation.
After some search I found out that there was a kind of DenseNet for semantic segmentation, so I replaced the old one with this. I hope it's right.
The official implementation is the one written with Lasagne.